Brought about by the use of extra compiler checks, notably by g95 and
the '-std=f2003 -fall-intrinsics' setting in gfortran.

Apart from syntax and visibility issues, F2008 extensions
('contiguous' keyword), and non-standard declarations (such as
"real*8") have been removed. Bounds specification in pointer
assignment is in principle permitted by F2003, but are not yet
implemented by all compilers.

Also, the building of the MPI interfaces has been updated to properly
include 8-byte integers.

(Declarations in the code in Util are not fully updated yet)

+++ Src/basis_specs.f
@@ -1067,9 +1067,9 @@ c (according to atmass subroutine).
       subroutine semicore_check(is)
       integer, intent(in)  :: is
 
-      real*8, parameter :: tiny = 1.d-5
+      real(dp), parameter :: tiny = 1.d-5
       integer ndiff
-      real*8 zval, zval_vps, charge_loc
+      real(dp) zval, zval_vps, charge_loc
 
       basp => basis_parameters(is)
